+++ src/Workspaces/Core/Portable/Classification/ClassifiedText.cs
@@ -1,0 +1,16 @@
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+
+namespace Microsoft.CodeAnalysis.Classification
+{
+    internal struct ClassifiedText
+    {
+        public string ClassificationType { get; }
+        public string Text { get; }
+
+        public ClassifiedText(string classificationType, string text)
+        {
+            ClassificationType = classificationType;
+            Text = text;
+        }
+    }
+}
